Notizerstellung nun funktionstüchtig
Notizen können nun auch über Client --> Logik --> DB erstellt werden,
habe dazu auch ein paar kleine Änderungen in der DB und am NoteMapper
vorgenommen @cbehr @lisakuechler
SQL File mit allen Änderungen im nächsten Commit

+++ src/de/hdm_stuttgart/huber/itprojekt/server/EditorImpl.java
@@ -73,7 +73,7 @@ public class EditorImpl extends RemoteServiceServlet implements Editor {
 		try {
 			
 			noteMapper.create(newNote);
-			newNote = noteMapper.findById(newNote.getNoteId());
+			newNote = noteMapper.findById((long) newNote.getNoteId());
 			
 		} catch (Exception e) {
 			e.printStackTrace();
@@ -91,7 +91,7 @@ public class EditorImpl extends RemoteServiceServlet implements Editor {
 		try { 
 			
 			noteMapper.save(newNote);
-			newNote = noteMapper.findById(newNote.getNoteId());
+			newNote = noteMapper.findById((long) newNote.getNoteId());
 			
 		} catch (Exception e) {
 			e.printStackTrace();
@@ -103,7 +103,7 @@ public class EditorImpl extends RemoteServiceServlet implements Editor {
 	@Override
 	public Note getNoteById(Note note) throws Exception {
 	
-			return noteMapper.findById(note.getNoteId());
+			return noteMapper.findById((long) note.getNoteId());
 	
 	}
